pyproject.toml
[project]
name = "sql-ai-datathon-python"
version = "0.1.0"
description = "RAG pipeline using PostgreSQL + pgvector and OpenAI — Microsoft SQL AI Datathon submission"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.131.0",
    "numpy>=2.4.2",
    "openai>=2.21.0",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.41.0",
]
